Fix the remaining compiler warnings.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@35044 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSAffineTransform.m
@@ -668,7 +668,7 @@ static NSAffineTransformStruct identityTransform = {
           else
             {
               // FIXME
-              NSLog(@"Got data %@ len %d for affine transform", d, length);
+              NSLog(@"Got data %@ len %d for affine transform", d, (int)length);
               return [self notImplemented: _cmd];
             }
         }
